Statement React component was born
