+++ package.json
@@ -18,7 +18,11 @@
   "homepage": "https://github.com/kib357/play-statement-ts#readme",
   "dependencies": {
     "@types/jest": "^27.0.1",
+    "@types/react": "^17.0.37",
+    "@types/react-test-renderer": "^17.0.1",
     "jest": "^27.0.6",
+    "react": "^17.0.2",
+    "react-test-renderer": "^17.0.2",
     "ts-jest": "^27.0.5",
     "typescript": "^4.3.5"
   }
